Event listing coloured background based on calendar color
